+++ packages/runtime/src/interpreter/opcode.ts
@@ -172,12 +172,12 @@ export class Op {
 
   /**
    * Returns substring from given string (if it exists)
+   * @param byteString given string as bytes
    * @param start starting index
    * @param end ending index
-   * @param byteString given string as bytes
    * @param line line number in TEAL file
    */
-  subString (start: bigint, end: bigint, byteString: Uint8Array, line: number): Uint8Array {
+  subString (byteString: Uint8Array, start: bigint, end: bigint, line: number): Uint8Array {
     if (end < start) {
       throw new RuntimeError(RUNTIME_ERRORS.TEAL.SUBSTRING_END_BEFORE_START, { line: line });
     }