+++ CHANGELOG.md
@@ -8,6 +8,10 @@
     * Loads, Stores opcodes
     * Extract, Extract3 opcodes
     * ExtractUint16, ExtractUint32, ExtractUint64 opcodes
+    * Txn, Global fields
+
+### Bug Fixes
++  Fix bug substring3 opcode pop wrong order [/#505](https://github.com/scale-it/algo-builder/pull/505), contribution: @vuvth.
 
 ## v2.1.0 2021-10-22